Extract createAgent*
